muneem-dev/requirements.txt
--extra-index-url https://__token__:8zKkkbRtGyJ1LdxxUshA@gitlab.renewbuy.in/api/v4/projects/103/packages/pypi/simple
--extra-index-url https://__token__:8zKkkbRtGyJ1LdxxUshA@gitlab.renewbuy.in/api/v4/projects/160/packages/pypi/simple
rb-utils==2023.2.17.7.49.53
integration-library==2023.5.25.10.57.29
aioboto3==10.1.0
aiofiles==0.8.0
aiobotocore==2.4.0
aiohttp==3.8.1
aioitertools==0.11.0
aiokafka==0.7.2
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
asyncpg==0.25.0
attrs==22.1.0
backcall==0.2.0
black==22.6.0
boto3==1.24.59
botocore==1.27.59
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==39.0.0
databases==0.6.0
decorator==5.1.1
docstring-parser==0.12
executing==1.0.0
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
frozenlist==1.3.1
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.2
mypy-extensions==0.4.3
num2words==0.5.12
numpy==1.23.3
ormar==0.11.1
pandas==1.5.0
parso==0.8.3
pdfkit==1.0.0
pathspec==0.9.0
pexpect==4.8.0
piccolo==0.80.1
piccolo-admin==0.21.0
pandas==1.5.0
piccolo-api==0.38.0
pickleshare==0.7.5
platformdirs==2.5.2
priority==2.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.1
Pygments==2.13.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.2.1
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
sqladmin==0.1.11
SQLAlchemy==1.4.37
stack-data==0.5.0
starlette==0.19.1
targ==0.3.7
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
typing-extensions==4.2.0
urllib3==1.26.12
uvicorn==0.17.6
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.0
